Move mock order data from model module to mock-data module.
