+++ server/worker/mock-data.js
@@ -1,4 +1,5 @@
 'use strict';
+const _ = require('lodash');
 const appRoot = require('app-root-path');
 const debugMockData = require('debug')('order-manager:Metric:mockData');
 const moment = require('moment');
@@ -86,6 +87,79 @@ async function getMockClients() {
 };
 
 /**
+ * Mock orders for the given clients.
+ * @param{Client[]} clients - mock clients
+ */
+async function mockOrderData(clients) {
+  if (!yn(process.env.CREATE_MOCK_DATA)) {
+    return;
+  }
+  debugMockData('Order.mockData() - Begins');
+  if (clients.length === 0) {
+    return;
+  }
+  const adminUser = await app.models.EndUser.findOne({ where: { role: 'admin' } });
+  const metadata = { endUserId: adminUser.id };
+  let orders = await app.models.Order.find({
+    where: { and: [
+      { clientId: { inq: clients.map(c => c.id) } },
+      { status: { nin: ['Cancelled', 'Completed'] } }
+    ] }
+  });
+
+  if (orders.length === 0) {
+    try {
+      const products = await app.models.Product.find();
+      const createResult = await Promise.mapSeries(clients, async function(client) {
+        let orderItems = products.map(function(product) {
+          if (Math.random() < 0.5) {
+            return {
+              productId: product.id,
+              quantity: Math.floor(Math.random() * 10),
+              unitPrice: Number(product.unitPrice)
+            };
+          }
+        });
+        orderItems = _.compact(orderItems);
+        let subtotal = orderItems.reduce((a, c) => a += c.quantity * c.unitPrice, 0);
+        let fee = (client.feeType === 'Fixed') ? Number(client.feeValue) : subtotal * Number(client.feeValue) / 100.0;
+        fee = Number(fee.toFixed(2));
+        let orderData = {
+          clientId: client.id,
+          status: 'Submitted',
+          subtotal: subtotal,
+          fee: fee,
+          totalAmount: subtotal + fee,
+          note: 'Mock data'
+        };
+        return await app.models.Order.createNew(orderData, orderItems, metadata);
+      });
+      orders = await app.models.Order.find({
+        where: { id: { inq: createResult.map(result => result.orderId) } }
+      });
+      orders.forEach(function(order) {
+        debugMockData(`<Client[${order.clientId}]>: Created new order(id: ${order.id})`);
+      });
+    } catch (error) {
+      console.error(`Error while creating mock orders - ${error.message}`);
+    }
+    return;
+  }
+
+  // update order status. 'Shipped' order updates its status via different API.
+  let shipped = _.remove(orders, function(order) { return order.status === 'Shipped'; });
+  await Promise.each(orders, async function(order) {
+    if (order.status === 'Submitted') {
+      await order.updateAttribute('status', 'Processed');
+    } else if (order.status === 'Processed') {
+      await order.updateAttribute('status', 'Shipped');
+    }
+  });
+  await app.models.Order.completeOrders(shipped.map(function(order) { return order.id; }), metadata);
+  debugMockData('Order.mockData() - Ends');
+};
+
+/**
  *  create mock data using random number generator.
  * [NOTE] this function is called 4 times a day, at 8, 12, 18, 20.
  */
@@ -100,7 +174,7 @@ module.exports.create = async function(fireDate) {
 
   let clients = await getMockClients();
   if (fireDate.getDay() > 0) {  // skip Sunday
-    await app.models.Order.mockData(clients);
+    await mockOrderData(clients);
   } else {
     await app.models.Statement.mockData(clients);
   }
